20260507-01: feat: Enhance SprayWeight application with weight measurement functionality
- Added weight measurement capabilities by introducing a new TCP connection to a weight sensor.
- Updated SprayWeightContext to include weightTcpHost and weightTcpPort for configuration.
- Implemented WeightWorker to handle weight measurement events and communicate with gPlat.
- Introduced StartWeightEvent and FinishWeightEvent structures for event handling.
- Enhanced error handling and logging for weight measurement processes.
- Updated configuration file to include weight sensor IP address and port.
- Modified existing code to ensure compatibility with new weight measurement features.

+++ SprayWeight/include/weight.h
@@ -14,7 +14,7 @@ public:
 
 private:
 	// 从串口读取重量值，单位为0.01kg。
-	double ReadWeightCentiKg() const;
+	int ReadWeightCentiKg() const;
 
 	// 共享运行时上下文，由 main() 创建。
 	SprayWeightContext& ctx_;
